Create cross-tabulation summary tables for srvyr objects

This function creates a main summary table and multiple subtables based on the specified variables, and combines them into a single formatted cross-tabulation. Designed specifically for survey design objects. Useful for weighted crosstabs.

srvy_crosstab(
  data,
  main_vars,
  sub_vars,
  stat_option = "both",
  custom_stat_cat = NULL,
  custom_stat_cont = NULL,
  main_digits = 0,
  sub_digits = 0,
  keep_footnotes = FALSE
)

Arguments

data

A survey design object

main_vars

A character vector of variable names to include in the main table.

sub_vars

A character vector of variable names to create subtables for

stat_option

String specifying the statistics to display: "single", "both", or "custom". Default is both. Single will only show percentage for categorical variables and mean for continuous variables. Both will show percentage and n size for categorical variables and mean and standard deviation for continuous variables (with the latter variable in parentheses). Custom allows you to specify custom configurations as specified in the custom_stat_cat and custom_stat_cont parameters.

custom_stat_cat

Only used when stat_option = "custom". The options follow those available under the "statistic" argument in gtsummary's tbl_svysummary.

custom_stat_cont

Only used when stat_option = "custom". The options follow those available under the "statistic" argument in gtsummary's tbl_svysummary.

main_digits

Number of digits for primary statistics

sub_digits

Number of digits for secondary statistics

keep_footnotes

Logical; whether to keep footnotes in the final table

Value

A gt table object

Examples

if (FALSE) { # \dontrun{
  library(Rsurveytools)
  data <- data
  svy <- svydesign(data = data, id=~1, weight = wts)
  srvy_crosstab(svy, main_vars = c("q1", "q2", "q3"), sub_vars = c("educ", "ideo5"))
} # }